Images are no longer saved as base64, they are saved to path instead

+++ Services/eCommerceBase/eCommerceBase/eCommerceBase.Application/Handlers/ProductPhotoes/Commands/CreateProductPhotoCommand.cs
@@ -6,10 +6,11 @@ using eCommerceBase.Persistence.GenericRepository;
 using eCommerceBase.Insfrastructure.Utilities.Caching.Redis;
 using eCommerceBase.Application.Constants;
 using eCommerceBase.Application.Handlers.Mapper;
+using eCommerceBase.Application.Helpers;
 
 namespace eCommerceBase.Application.Handlers.ProductPhotoes.Commands;
 public record CreateProductPhotoCommand(Guid ProductId,
-		string PhotoBase64) : IRequest<Result>;
+		string ImageUrl) : IRequest<Result>;
 public class CreateProductPhotoCommandHandler(IWriteDbRepository<ProductPhoto> productPhotoRepository,
 		IUnitOfWork unitOfWork,
 		ICacheService cacheService) : IRequestHandler<CreateProductPhotoCommand,
@@ -23,7 +24,15 @@ public class CreateProductPhotoCommandHandler(IWriteDbRepository<ProductPhoto> p
     {
         return await _unitOfWork.BeginTransaction(async () =>
         {
+            var id = Guid.NewGuid();
             var data = ProductPhotoMapper.CreateProductPhotoCommandToProductPhoto(request);
+            data.SetId(id);
+            var savePhoto = await PhotoHelper.SaveBase64ImageAsWebP(id, request.ImageUrl);
+            if (!savePhoto.Success)
+            {
+                return Result.SuccessResult(savePhoto.Message);
+            }
+            data.SetImageUrl(savePhoto.Data!);
             await _productPhotoRepository.AddAsync(data);
             await _cacheService.RemovePatternAsync("eCommerceBase:Product");
             return Result.SuccessResult(Messages.Added);
